refactor: the writer stamps hint locks too, closing the last read-back
`lock_keys::apply` was the last thing writing map grid tiles, and its own
comment recorded the chain that forced it: stamp the hint tiles first,
because `removable_rows` reads the map to decide what the table needs.
That is why it was also the last module reading the finished map back off
the ROM instead of being handed it.

Both halves move. `overworld_writer::grid` picks the lock byte, hint
variants included — it already picked the plain one from the path
underneath, and after the pairing became a `FortRef` the model can say
"this key is in another world" (`lock.fort.world != wi`). The vocabulary
stays in `lock_keys` behind `lock_tile` / `local_sky_tile`, beside the
metatile art and the removable pairing that have to agree with it.

The local-sky rule stays a **sweep, not a per-lock branch**, because that
is what it is: sky's alternate colour is its plain tile, so a remote sky
lock changes no byte and a rule keyed on the tile alone would sweep the
remote ones up with the local ones. The writer holds the whole grid, so it
reproduces the sweep exactly.

`lock_keys::apply` then takes the grids and writes no map tile:
`define_hint_metatiles` composes art for the hint indices actually present
(per tile now, not per lock — several locks sharing one wrote identical
bytes before), `tiles_on_map` and `removable_rows` read the grids, and the
away-lock bit lookup uses `CompletionMap::from_grids`. `hammer_breaks`
takes them too, for the same reason.

With that, **nothing writes a map grid tile after the overworld writer**:
`qol`'s map patches run before the builder, and `wand_gate` and
`lock_keys` are test-only there now. `CompletionMap::from_rom` and
`read_all_tile_grids` survive only for tests and `testrom`, which patches
a finished ROM with no writer in the path — both are gated to say so, and
the wasm clippy pass is what noticed the second one.

`overworld_writer/fortress_fx.rs` retires with it: one live function, now
inlined into `lock_entries`, in a module named for the subsystem
`lock_keys` replaced. Its tests move to the writer's own — they pin
vanilla tables `overworld_pickup::open_fx_gaps` still reads.

Byte-identical, standard and maze, all 92 outputs.
